# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Todoch API",
    "version": "0.0.8"
  },
  "paths": {
...
    "/api/events": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "List Events",
        "operationId": "list_events_api_events_get",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "To"
            }
          },
          {
            "name": "area_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Area Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OccurrenceOut"
                  },
                  "title": "Response List Events Api Events Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "events"
        ],
        "summary": "Create Event",
        "operationId": "create_event_api_events_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventWriteOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/events/{event_id}": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Get Event",
        "operationId": "get_event_api_events__event_id__get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Event Id"
            }
          },
          {
            "name": "occurrence",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Occurrence"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "events"
        ],
        "summary": "Update Event",
        "operationId": "update_event_api_events__event_id__patch",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Event Id"
            }
          },
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "this",
                "following",
                "all"
              ],
              "type": "string",
              "default": "all",
              "title": "Scope"
            }
          },
          {
            "name": "occurrence",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Occurrence"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventWriteOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "events"
        ],
        "summary": "Delete Event",
        "operationId": "delete_event_api_events__event_id__delete",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Event Id"
            }
          },
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "this",
                "following",
                "all"
              ],
              "type": "string",
              "default": "all",
              "title": "Scope"
            }
          },
          {
            "name": "occurrence",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Occurrence"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/events/{event_id}/ics": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Event Ics",
        "description": "Einzelner Termin (bei Serien mit allen Ausnahmen) als .ics-Datei zum Weitergeben.\n\nStandard ist ``public``: ohne Beschreibung, damit interne Gesprächsnotizen nicht beim\nGegenüber landen.",
        "operationId": "event_ics_api_events__event_id__ics_get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Event Id"
            }
          },
          {
            "name": "detail",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "full",
                "public"
              ],
              "type": "string",
              "default": "public",
              "title": "Detail"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EventIn": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Title"
          },
          "description": {
            "type": "string",
            "maxLength": 50000,
            "title": "Description",
            "default": ""
          },
          "location": {
            "type": "string",
            "maxLength": 500,
            "title": "Location",
            "default": ""
          },
          "url": {
            "type": "string",
            "maxLength": 1000,
            "title": "Url",
            "default": ""
          },
          "area_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area Id"
          },
          "all_day": {
            "type": "boolean",
            "title": "All Day",
            "default": false
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "start_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Time"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "end_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "time"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Time"
          },
          "rrule": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Rrule"
          },
          "status": {
            "type": "string",
            "enum": [
              "tentative",
              "confirmed",
              "cancelled"
            ],
            "title": "Status",
            "default": "confirmed"
          },
          "transparency": {
            "type": "string",
            "enum": [
              "opaque",
              "transparent"
            ],
            "title": "Transparency",
            "default": "opaque"
          },
          "is_fixed": {
            "type": "boolean",
            "title": "Is Fixed",
            "default": false
          },
          "tags": {
            "items": {
              "type": "string",
              "maxLength": 40,
              "minLength": 1,
              "pattern": "^[\\w-]+$"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Tags",
            "default": []
          },
          "attendees": {
            "items": {
              "$ref": "#/components/schemas/AttendeeIn"
            },
            "type": "array",
            "maxItems": 50,
            "title": "Attendees",
            "default": []
          },
          "reminders": {
            "items": {
              "type": "integer",
              "maximum": 40320.0,
              "minimum": 0.0
            },
            "type": "array",
            "maxItems": 10,
            "title": "Reminders",
            "default": []
          },
          "tzid": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Tzid"
          },
          "contact_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Id"
          },
          "channel": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "phone",
                  "in_person",
                  "mail",
                  "other"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel"
          },
          "agreed_on": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agreed On"
          },
          "agreed_with": {
            "type": "string",
            "maxLength": 200,
            "title": "Agreed With",
            "default": ""
          },
          "priority": {
            "type": "integer",
            "maximum": 3.0,
            "minimum": 0.0,
            "title": "Priority",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "title",
          "start_date"
        ],
        "title": "EventIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OccurrenceOut": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "event_id": {
            "type": "string",
            "format": "uuid",
            "title": "Event Id"
          },
          "series_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Series Id"
          },
          "recurrence_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recurrence Id"
          },
          "area_id": {
            "type": "string",
            "format": "uuid",
            "title": "Area Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "all_day": {
            "type": "boolean",
            "title": "All Day"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "is_fixed": {
            "type": "boolean",
            "title": "Is Fixed"
          },
          "recurring": {
            "type": "boolean",
            "title": "Recurring"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "start": {
            "type": "string",
            "format": "date-time",
            "title": "Start"
          },
          "end": {
            "type": "string",
            "format": "date-time",
            "title": "End"
          },
          "start_local": {
            "type": "string",
            "title": "Start Local"
          },
          "end_local": {
            "type": "string",
            "title": "End Local"
          }
        },
        "type": "object",
        "required": [
          "key",
          "event_id",
          "series_id",
          "recurrence_id",
          "area_id",
          "title",
          "location",
          "all_day",
          "status",
          "is_fixed",
          "recurring",
          "tags",
          "start",
          "end",
          "start_local",
          "end_local"
        ],
        "title": "OccurrenceOut"
      },
...
    }
  }
}